% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellGroupSummary.R
\name{cellGroupSummary}
\alias{cellGroupSummary}
\title{Produces summary of cell groups from fit objects}
\usage{
cellGroupSummary(fits, cellsToUse = NULL, highConfOnly = FALSE)
}
\arguments{
\item{fits}{A list containing fit objects produced by \code{\link{inferInactiveX}}, giving Xi states from multiple individuals.}

\item{cellsToUse}{If NULL, all cells.  If a vector of cell IDs, pop size for that cell type.  If a list of cellIDs, pop size for all.}

\item{highConfOnly}{Use all estimates of Xi, or just the high confidence calls?}
}
\value{
A data.table giving the Xi summary stats by cell group and individual from fit objects.
}
\description{
Produces table summarising the Xi skew for each cell type in \code{cellsToUse} for each individual in \code{fits}.  Individiual and cell type names are taken from the names of \code{cellsToUse} and \code{fits} if present.
}
