% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministicAnnealing.R
\name{deterministicAnnealing}
\alias{deterministicAnnealing}
\title{Deterministic annealing expectation maximisation}
\usage{
deterministicAnnealing(
  dd,
  betaStart,
  betaFac,
  tauInit,
  anchorCell,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dd}{A data.frame where rows are observations at a particular cell/SNP combination.  Columns must be \code{refCount}, \code{altCount}, \code{tot}, \code{cell}, and \code{SNP}.}

\item{betaStart}{Initial beta to use for deterministic annealing.}

\item{betaFac}{How much to increase beta by in each pass through the deterministic annealing loop.}

\item{tauInit}{Initial guess for fraction of cells expressing maternal allele.}

\item{anchorCell}{ID of cell assumed to be maternally active.  If NULL, one with highest coverage used.}

\item{verbose}{Be verbose?  Levels are 0 (off), 1 (outer loop), 2 (beta loop), 3 (EM loop)}

\item{...}{Extra parameters passed to \code{\link{inactiveXEM}}}
}
\value{
Fit summary for
}
\description{
Uses the deterministic annealing approach to wrap around \code{\link{inactiveXEM}}.  This smooths the likelihood function strongly to begin with, then successively reduces the smoothing in increments until it is zero.  This approach should allow the EM algorithm a better chance of converging to a global, rather than local, maximum.
}
\details{
Not intended for direct use, call \code{\link{inferInactiveX}} instead.
}
