% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCorrelation.R
\name{estCorrelation}
\alias{estCorrelation}
\title{Estimate correlation of Xi skew between cell groups}
\usage{
estCorrelation(
  fits,
  cellsToUse,
  groupsToCompare = names(cellsToUse),
  relativeTo = c("fixed", "median", "tau"),
  fixedVal = 0,
  exactConf = FALSE,
  highConfOnly = FALSE,
  priorAlpha = 1,
  priorBeta = 1,
  nSamps = 1000,
  distQuants = c(0.01, 0.05, 0.1, 0.9, 0.95, 0.99),
  verbose = 0,
  nParallel = 1,
  mcParams = list()
)
}
\arguments{
\item{fits}{A list containing fit objects produced by \code{\link{inferInactiveX}}, giving Xi states from multiple individuals.}

\item{cellsToUse}{List containing at least two entries, named by cell types, giving the cellIDs for the cells of that type.}

\item{groupsToCompare}{Which of the cell groupings defined by `cellsToUse` to calculate correlation on.}

\item{relativeTo}{Should we calculate correlation on raw Xi skew, or Xi skew relative to something?}

\item{fixedVal}{Value to measure Xi skew differences from.  Ignored unless `relativeTo='fixed'`.}

\item{exactConf}{Should the exact distribution be used to calculate uncertainty (slower, but more accurate)?}

\item{highConfOnly}{Use all estimates of Xi, or just the high confidence calls?}

\item{priorAlpha}{alpha pramater for beta distribution prior on individual Xi skew.}

\item{priorBeta}{beta pramater for beta distribution prior on individual Xi skew.}

\item{nSamps}{Number of random samples used to estimate uncertainty.}

\item{distQuants}{Which quantiles of the posterior distribution to provide?}

\item{verbose}{Level of verbosity, higher numbers is more.}

\item{nParallel}{How many threads in parallel}

\item{mcParams}{Overwrite default parameters controlling the call to \code{\link{mclapply}}.  Only used when \code{exactConf=TRUE} and \code{nParallel>1}}
}
\value{
A data.frame giving point estimates and confidence distributions for each pairwise comparison.
}
\description{
Given Xi skew estimates from a range of individuals, calculate the degree to which different groups of cells Xi skew are correlated.  This correlation can either be absolute, or relative to some estimate of the founder population skew (usually the mean or median Xi skew across cell types).
}
\details{
As with \code{\link{estPopSize}}, there are two sources of uncertainty in these estimates, which depend on the number of cells of a given type within individuals, and the total number of individuals.  Both are taken into account in this estimation.

Correlation coefficients are calculated for all pairwise combinations of cell types in \code{cellsToUse} that are listed in \code{groupsToCompare}.  

When calculating correlation based on raw Xi skew, there is no need to include cell groupings other than the ones you are interested in in \code{cellsToUse}.  However, when \code{relativeTo} is set to something other than 'none', other cells types are needed.  This is because the comparison point is itself a derived quantity from the Xi skews, and as such should be derived independently from each random sample from the posterior. 

Options for code{relativeTo} are: 'fixed', which correlates the Xi skew relative to \code{fixedVal} (0 by default), 'median', which correlates the difference of each Xi skew from the median value across all cell groupings, and 'tau', which correlates the difference from the average across all cells.

When \code{exactConf} is \code{TRUE}, the exact sampling distribution for the correlation coefficient is used to estimate uncertainty due to number of individuals.  This is more accurate, particularly for low numbers of individuals, but slow (usually not so slow as to be a major issue).  When set to \code{FALSE}, the Fisher transformation is instead used.
}
