% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPopSize.R
\name{estPopSize}
\alias{estPopSize}
\title{Estimates effective population bottleneck size}
\usage{
estPopSize(
  fits,
  cellsToUse = NULL,
  highConfOnly = FALSE,
  priorAlpha = 1,
  priorBeta = 1,
  nSamps = 1000,
  distQuants = c(0.01, 0.05, 0.1, 0.9, 0.95, 0.99)
)
}
\arguments{
\item{fits}{A list containing fit objects produced by \code{\link{inferInactiveX}}, giving Xi states from multiple individuals.}

\item{cellsToUse}{If NULL, all cells.  If a vector of cell IDs, pop size for that cell type.  If a list of cellIDs, pop size for all.}

\item{highConfOnly}{Use all estimates of Xi, or just the high confidence calls?}

\item{priorAlpha}{alpha pramater for beta distribution prior on individual Xi skew.}

\item{priorBeta}{beta pramater for beta distribution prior on individual Xi skew.}

\item{nSamps}{Number of random samples used to estimate uncertainty.}

\item{distQuants}{Which quantiles of the posterior distribution to provide?}
}
\value{
A data.frame containing the estimate of the populaion size with uncertainty.
}
\description{
Given Xi data from the same group of cells (cell type, cell state, tissue, etc.), estimate the size of the population bottleneck that would be needed to produce the observed spread in Xi skew across individuals.  This value will represent the smallest population bottleneck in the ancestory of the cell group investigated, which for many cells will be the "founder cell number"; the number of cells present when X-inactivation was established giving rise to this lineage.
}
\details{
The estimation is performed using the model that the population size is related to the variance of the Xi skew across individuals as N = 1/(4*var), which is derived from the assumption that Xi skew for an individual is binomially distributed across the founder cells (with p=0.5).  This function takes into account two sources of uncertainty in estimating the effective population bottleneck size N.  Firstly, in each individual, the measurement of Xi skew has uncertainty, which decreases as the number of cells in each individual increases.  Secondly, the variance of the distribution of Xi skew from a finite number of individuals is uncertain.  Typically, most of the unceratainty is driven by the number of individuals, not the number of cells in each individual.

To estimate the uncertainty on the measured Xi skew in each individual, a beta prior with parameters alpha and beta are used.  The default values imply a completely uniform prior.

By default, this function will calculate the population size for one group of cells.  Multiple groups can be estimated simultaneously if \code{cellsToUse} is given as a list of cellIDs, where then one estimate will be produced per list entry and named by the names of the \code{cellsToUse} list.
}
