% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCountsX.R
\name{filterCountsX}
\alias{filterCountsX}
\title{Filter X counts}
\usage{
filterCountsX(
  cnts,
  cellsToUse = NULL,
  minObs = 2,
  dropBiallelic = TRUE,
  dropStrata = c("PAR1", "PAR2", "strata2", "strata3"),
  regionsToUse = c("Exonic", "Intronic"),
  errRate = 0.02,
  pCut = 0.1,
  autoSet = TRUE
)
}
\arguments{
\item{cnts}{The counts to filter.  Usually the output of \code{\link{hetSNPsFromRNA}}, \code{\link[alleleIntegrator]{getAllelicExpression}}, or \code{\link{vartrixCnts}}.}

\item{cellsToUse}{Filter out any cells not in this list.  If NULL, no filtering on cell barcdodes.}

\item{minObs}{Drop any SNP that is not observed in at least this many cells.}

\item{dropBiallelic}{Drop any SNP that has evidence of escaping X-inactivation.}

\item{dropStrata}{Drop those in these stratas.  Usually these mark regions where the chance of escape of X-Inactivation is high.}

\item{regionsToUse}{Drop regions not off this type.}

\item{errRate}{Error rate to assume for finding biallelic SNPs}

\item{pCut}{P-value cut to use for finding biallelic SNPs.}

\item{autoSet}{If strata and/or regions not set, set them automatically using a GRCh38 reference.}
}
\value{
Filtered counts.
}
\description{
Given allele and cell resolved, counts at heterozygous SNPs on X, filter them down to just the useful ones.  \code{errRate} and \code{pCut} have both been calibrated quite precisely, don't change them unless you know what you're doing.
}
\details{
In some cases, it may be useful to set \code{dropBiallelic} to \code{FALSE}.  The logic behind dropping biallelic SNPs (that is those for which both alleles are detected in the same cell) is sound.  However, for SNPs with high coverage, this can occur just due to errors a non-trivial fraction of the time.  A simple statistical test is used to try and reduce this, but will not be perfect.  While discarding SNPs with evidence of both alleles is still the consernative thing to do, it may come at the cost of throwing out too much data.

Usually the output of \code{\link{hetSNPsFromRNA}} will be the input for this function (to \code{cnts}).  However, if for some reason you have calculated cell specific allele counts (usually because you have supplied a list of known heterozygous SNPs) from running \code{\link[alleleIntegrator]{getAllelicExpression}} or similar, that can also be supplied as input.

Filtering is done on both the evolutionary strata (excluding those known to escape X-inactivation) and the SNPs location relative to genes (e.g. intronic, exonic, or intergenic).  To do this, it is assumed that a column called \code{regionType} will encode the SNPs relationship to genes and \code{strata} will encode the evolutionary strata.  This are pre-calculated when running \code{\link{hetSNPsFromRNA}} with default parameters, but if other inputs are used they will not be present.  In this case the user may either manually create these collumns, or (assuming \code{autoSet==TRUE}) allow them to be automatically created assuming a GRCh38 reference from the \code{\link{X1k}} data object.
}
