% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellBarcodes.R
\name{getCellBarcodes}
\alias{getCellBarcodes}
\title{Get valid cell barcodes for 10X experiments}
\usage{
getCellBarcodes(
  dataPaths,
  checkSex = TRUE,
  genesX = "XIST",
  genesY = c("RPS4Y1", "RPS4Y2")
)
}
\arguments{
\item{dataPaths}{Named vector giving paths containing cellranger filtered cells to load.  Names must be unique giving the name to use for each 10X channel.}

\item{checkSex}{Should we check if all samples contain only X chromosomes?}

\item{genesX}{Genes specific to X chromosome, used to check sex.}

\item{genesY}{Genes specific to Y chromosome, used to check sex.}
}
\value{
A vector of cell barcodes to be passed to the \code{cellsToUse} param of \code{\link{hetSNPsFromRNA}}
}
\description{
Given a set of cellranger outputs from which the X-inactivation status is to be estimated, fetches the barcodes considered to have cells.  You can either specify paths as you would when running \code{\link{Read10X}} (which is what this function uses internally), or pass the \code{BAMs} argument to be used by \code{\link{hetSNPsFromRNA}} and the path will be guessed.
}
\details{
Note that \code{dataPaths} must be a named vector, as the names will be used to construct the allowable cell IDs.

As well as getting the valid barcodes, this function will also do a basic check to ensure that the samples provided look like they only contain X chromosomes.  This is based on determining if more cells express at least one of the \code{genesX} or \code{genesY}.
}
