% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetSNPsFromRNA.R
\name{hetSNPsFromRNA}
\alias{hetSNPsFromRNA}
\title{Infer heterozygous SNPs from scRNA}
\usage{
hetSNPsFromRNA(
  BAMs,
  refGenome,
  SNPs = X1k,
  cellsToUse = NULL,
  errRate = 0.05,
  alpha = 0.05,
  preserve = c("REF", "ALT", "AF", "strata", "regionType", "geneID", "geneName"),
  useVartrix = FALSE,
  ...
)
}
\arguments{
\item{BAMs}{The 10X BAMs to use.}

\item{refGenome}{The reference genome used to map \code{BAMs}.}

\item{SNPs}{Which SNPs to consider.  Default are 1k genome SNPs on X with frequency greater than 1\%.  Note these have been lifted over to the GRCh38 reference.}

\item{cellsToUse}{Only look at cells with these barcodes.  Must be specified if using vartrix.}

\item{errRate}{When testing for presence of allele, use this error rate.}

\item{alpha}{Consider an allele detected at this significance threshold.}

\item{preserve}{Which metadata entries to preserve (if they're present in SNPs)}

\item{useVartrix}{Should we use vartrix to get allele counts?}

\item{...}{Passed to \code{\link{getAllelicExpression}}}
}
\value{
A GRanges object with the heterozygous SNPs.  Stored in the @metadata slot is also the cell specific counts for those snps to save future calculation.
}
\description{
Looks at locations of common SNPs in the human population (on chromosome X by default) and keeps as heterozygous any location where both alleles are detected in the data.  Optionally only consider those cells in \code{cellsToUse} when doing this.
}
\details{
If there are multiple 10X BAM files relating to the same individual, all should be passed to this function (via the \code{BAMs} parameter) at the same time to maximise the power to detect both alleles.  This depends on the alleleCount binary being present, or the vartrix binary if \code{useVartrix} is set.

As well as returning locations of heterozygous SNPs, this function also stores the raw counts at each cell generated as part of the estimation procedure.  These can then be reused for estimation of the X inactivation state by passing the output of \code{\link{hetSNPsFromRNA}} to \code{\link{filterCountsX}} (see help for \code{\link{filterCountsX}} for more details).
}
