% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactiveXEM.R
\name{inactiveXEM}
\alias{inactiveXEM}
\title{Inner EM loop}
\usage{
inactiveXEM(
  dd,
  beta,
  tauInit,
  lambdas,
  states,
  errRate,
  tol,
  maxIter,
  verbose = FALSE
)
}
\arguments{
\item{dd}{A data.frame where rows are observations at a particular cell/SNP combination.  Columns must be \code{refCount}, \code{altCount}, \code{tot}, \code{cell}, and \code{SNP}.}

\item{beta}{Deterministic annealing parameter for this fit. beta=1 solves the unperterbed likelihood function.}

\item{tauInit}{Initial guess for fraction of cells expressing maternal allele.}

\item{lambdas}{SNP state.  Maternal allele is alt when 1 and ref when 0.}

\item{states}{Cell states.  Maternal genotype when 1, (i.e. paternal X inactivated)}

\item{errRate}{The error rate to asume.}

\item{tol}{EM terminated when change in Q less than this value.}

\item{maxIter}{EM terminated after this many iterations, regardless of change in Q.}

\item{verbose}{Be verbose?  Levels are 0 (off), 1 (outer loop), 2 (beta loop), 3 (EM loop)}
}
\value{
List of states, genotype, tau, iteration number, Q, dQ
}
\description{
Runs the EM loop to estimate the X chromosome genotype and inactivation state.  Not intended to be used directly.
}
