% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactiveXX.R
\name{inactiveXX}
\alias{inactiveXX}
\title{Infer X inactivation state for all cells}
\usage{
inactiveXX(bams10X, refGenome, verbose = TRUE)
}
\arguments{
\item{bams10X}{BAM files output by cellranger.  Should include every cellranger BAM file relating to one individual with at least two X chromosomes.}

\item{refGenome}{The reference genome used to map the BAMs provided.}

\item{verbose}{Be verbose?}
}
\value{
The output of \code{\link{inferInactiveX}}.
}
\description{
This is a wrapper script that will perform the standard workflow to estimate the X-Inactivation state for every cell in a 10X experiment from the BAM files.
}
