% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferInactiveX.R
\name{inferInactiveX}
\alias{inferInactiveX}
\title{Infer which copy of X is inactivated}
\usage{
inferInactiveX(
  cnts,
  errRate = 0.1,
  logitCut = 3,
  pCut = 0.2,
  tauInit = 0.5,
  betaStart = 0.01,
  betaFac = 1.3,
  anchorCell = NULL,
  nStarts = 1000,
  tol = 1e-06,
  maxIter = 1000,
  nParallel = 1,
  verbose = 1
)
}
\arguments{
\item{cnts}{Usually the output of \code{\link{getAllelicExpression}}.  A GRanges of counts with refCount/altCount (matCount/patCount if pre-phased) with one row for each cell and SNP.}

\item{errRate}{Assumed all cause error rate.  10\% seems to be about the baseline, but can vary}

\item{logitCut}{Cells with probabilities (on logit) scale greater than this value are marked high confidence cells.}

\item{pCut}{Cells with p-value of excess discrepent counts under this value cannot be high confidence cells.}

\item{tauInit}{Initial guess for fraction of cells expressing maternal allele.}

\item{betaStart}{Initial beta to use for deterministic annealing.}

\item{betaFac}{How much to increase beta by in each pass through the deterministic annealing loop.}

\item{anchorCell}{ID of cell assumed to be maternally active.  If NULL, one with highest coverage used.}

\item{nStarts}{How many times to run the fit starting at random locations.}

\item{tol}{EM terminated when change in Q less than this value.}

\item{maxIter}{Terminate EM if more than this many iterations.}

\item{nParallel}{How many threads?}

\item{verbose}{Be verbose?  Levels are 0 (off), 1 (outer loop), 2 (beta loop), 3 (EM loop)}
}
\value{
A list with \code{states}, indicating which allele is active, \code{genotype} indicating if the reference allele is maternal for each SNP, \code{tau} the fraction of cells with maternal active, \code{allFits} which contains results for each of the \code{nStarts} random initiations, \code{cellSummary} which contains summary stats for each cell, \code{snpSummary} which contains summary stats for each SNP, and \code{dd} which contains the raw data.
}
\description{
Given counts on the X chromosome, use EM to simultaneously estimate the genotype of the X chromosome and which version is inactivated in each cell.  If \code{cnts} includes columns \code{matCount} and \code{patCount}, the genotypes are assumed to have already been phased and rather than estimating the genotype, the genotype is assumed.
}
\details{
To ensure consistency of results, one cell (the \code{anchorCell}) is assumed to be in the maternally active state.  If not specified, this is picked to be the cell with the highest X chromosome coverage.

The summary stats provided at the cell/SNP level essentially assume the other part of the data is correct and then measure how discrepent each cell or SNP is.  For example, the cell summary stats assumes all SNPs have been correctly phased and then measures how many reads are in conflict with this phasing across each cell.
}
