% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{plotCorrelations}
\alias{plotCorrelations}
\title{Plot Xi skew correlations}
\usage{
plotCorrelations(rhos, nBins = 50, ...)
}
\arguments{
\item{rhos}{Output of \code{\link{estCorrelation}}.}

\item{nBins}{How many bins to use for histograms showing uncertainty distribution.}

\item{...}{Passed to \code{\link{Heatmap}}.}
}
\value{
A heatmap object.
}
\description{
Visualise the output of \code{\link{estCorrelation}} as a heatmap.  The organisation of the heatmap is driven by the point estimates of the correlation coefficient for each pair of cell groupings, while each tile shows the distribution over the uncertainty in those estimates.
}
\details{
The output is a \code{ComplexHeatmap} object and so can be controlled in the normal way for that package.

If you encounter errors about NAs, try disabling hierarchical clustering by passing \code{cluster_rows=FALSE} and \code{cluster_columns=FALSE} to \code{\link[ComplexHeatmap]{Heatmap}}.
}
