% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelHeatmap.R
\name{plotModelHeatmap}
\alias{plotModelHeatmap}
\title{Plot model heatmap}
\usage{
plotModelHeatmap(
  fit,
  summariseBy = c("none", "SNP", "cell"),
  tol = 0.1,
  minCov = 0,
  badCutoff = 0.1,
  truncAt = 10
)
}
\arguments{
\item{fit}{The output of \code{\link{inferInactiveX}}.}

\item{summariseBy}{Should the heatmap summarise by SNP, cell, or do no manipulation of data (the default).}

\item{tol}{Binarise cells into maternal/paternal with this tolerance.}

\item{minCov}{Only plot those SNPs with at least this many reads.}

\item{badCutoff}{Any SNP where the 95\% confidence interval for counts going against the model exceeds this fraction is marked as "bad".}

\item{truncAt}{Truncate logit probabilities with absolute value above this.}
}
\value{
Nothing.  A heatmap is drawn.
}
\description{
Make a plot summarising the model.  Specifically plots a heatmap of SNPs by cells.  This is annotated by various summary metrics for each SNP and cell, which are calculated assuming the other data modality is true (i.e., number of bad reads at each SNP calculated assuming the cell assignment is correct).
}
\details{
By default, the heatmap shows the (usually very sparse) raw data.  However, if \code{summariseBy} is set to \code{"SNP"} or \code{"cell"}, the heatmap is distorted to remove the sparsity and emphasise the performance of either the cell or SNP (depending on the value of \code{summariseBy}).

When summarising by SNP (or cell), all NA values are removed from each row (or column).  This allows visualisation of the overall data for each SNP (or cell), at the cost of making comparisons between rows (or columns) more difficult.
}
