% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSolutions.R
\name{plotSolutions}
\alias{plotSolutions}
\title{Plot summary of solutions}
\usage{
plotSolutions(fit, jitter = c(0.01, 0.01))
}
\arguments{
\item{fit}{The output of \code{\link{inferInactiveX}}.}

\item{jitter}{Fraction of plot range to jittery by in x,y direction, as a fraction of the dynamic range.}
}
\value{
Nothing, but a plot is produced.
}
\description{
The EM algorithm used by \code{\link{inferInactiveX}} is rerun with many random starting points (1,000 by default).  This function plots a summary of these solutions, showing the log likelihood of each solution (y-axis) against the fraction of cells allocated to maternal (x-axis).
}
\details{
When the genotyping is not known beforehand the designation of maternal or paternal is arbitrary and you should expect the solutions to be evenly distributed either side of 0.5 on the x-axis.

Jitter is added to points to allow the high density regions of the solution space to be visualised.
}
