% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.inactiveXX}
\alias{print.inactiveXX}
\title{Print summary of fit object}
\usage{
\method{print}{inactiveXX}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Unused.}
}
\value{
Nothing.  Print a message.
}
\description{
Summary of features of fit returned by \code{\link{inferInactiveX}}.
}
