% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleExactConfRho.R
\name{sampleExactConfRho}
\alias{sampleExactConfRho}
\title{Sample from exact sampling distribution for pearson correlation}
\usage{
sampleExactConfRho(r, n, nSamps = 1000)
}
\arguments{
\item{r}{Observed pearson correlation.}

\item{n}{Number of paired observations from which r was derived.}

\item{nSamps}{Number of random samples to draw.}
}
\value{
\code{nSamps} from the sampling distribution of the correlation coefficient.
}
\description{
Given a pearson correlation coefficient \code{r} estimated from \code{n} observations, randomly sample \code{nSamps} from the sampling distribution.
}
\details{
The standard \code{\link{cor.test}} simply uses Fisher's transformation to approximate the sampling distribution.  But this breaks down for small n or strong correlation, exactly where the uncertainty matters.  This function implements the exact sampling distribution pdf.  As such, it will obviously be slower, but more accurate.
}
