% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartrixCnts.R
\name{vartrixCnts}
\alias{vartrixCnts}
\title{Get counts at variants using vartrix}
\usage{
vartrixCnts(
  loci,
  BAMs,
  barcodes,
  refGenome,
  outputs = NULL,
  skipIfExists = TRUE,
  mapq = 30,
  nBasesPad = 500,
  autoChr = TRUE,
  nParallel = 1,
  noOutputCheck = FALSE,
  ...
)
}
\arguments{
\item{loci}{GRanges object that contains REF and ALT.}

\item{BAMs}{Cellranger 10X BAM files, or at least the least something that has the cellranger tags for UMI and CB.  Must be uniquely named by a label to form cell IDs.}

\item{barcodes}{The barcodes (or barcode files) to consider.}

\item{refGenome}{fasta reference genome to use.}

\item{outputs}{Path to save output for each BAM to.  Temp files used if NULL.}

\item{skipIfExists}{Should we load from file if possible.}

\item{mapq}{mapq threshold to use.}

\item{nBasesPad}{How many bases on either side of variant to use in mapping.}

\item{autoChr}{Automatically try and strip/add 'chr' to chromosome names as needed.}

\item{nParallel}{How many threads.}

\item{noOutputCheck}{Suppresses output warning, used for recursion.}

\item{...}{Currently does nothing.}
}
\value{
Loci, expanded to have one entry per cell/snp combination with reads, with refCount and altCount columns added.
}
\description{
Rather than using alleleCounter, with all the cruft that requires, use the dedicated 10X tool vartrix to get counts.  Unlike alleleCount, it will work with indels and other things just as well as single base pair events.
}
\details{
Only BAMs and barcodes are allowed to be not length 1, in which case their lengths must match (if files) or barcodes must contain cellID_barcode formatted barcodes.   All BAMs are run with the same parameters in series.  The results are made non-ambiguous by the cellID, which is a concatenation of the BAM file label and the barcode, separated by an underscore.
}
