% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCountsX.R
\name{filterCountsX}
\alias{filterCountsX}
\title{Filter X counts}
\usage{
filterCountsX(
  cnts,
  cellsToUse = NULL,
  minObs = 2,
  dropBiallelic = TRUE,
  dropStrata = c("PAR1", "PAR2", "strata2", "strata3"),
  regionsToUse = c("Exonic", "Intronic"),
  errRate = 0.02,
  pCut = 0.1
)
}
\arguments{
\item{cnts}{The counts to filter.  Should be the output of either \code{\link{getAllelicExpression}}, \code{\link{vartrixCnts}}, or \code{\link{hetSNPsFromRNA}}.}

\item{cellsToUse}{Filter out any cells not in this list.  If NULL, no filtering on cell barcdodes.}

\item{minObs}{Drop any SNP that is not observed in at least this many cells.}

\item{dropBiallelic}{Drop any SNP that has evidence of escaping X-inactivation.}

\item{dropStrata}{Drop those in these stratas.  Usually these mark regions where the chance of escape of X-Inactivation is high.}

\item{regionsToUse}{Drop regions not off this type.}

\item{errRate}{Error rate to assume for finding biallelic SNPs}

\item{pCut}{P-value cut to use for finding biallelic SNPs.}
}
\value{
Filtered counts.
}
\description{
Given allele and cell resolved, counts at heterozygous SNPs on X, filter them down to just the useful ones.  \code{errRate} and \code{pCut} have both been calibrated quite precisely, don't change them unless you know what you're doing.
}
\details{
In some cases, it may be useful to set \code{dropBiallelic} to \code{FALSE}.  The logic behind dropping biallelic SNPs (that is those for which both alleles are detected in the same cell) is sound.  However, for SNPs with high coverage, this can occur just due to errors a non-trivial fraction of the time.  A simple statistical test is used to try and reduce this, but will not be perfect.  While discarding SNPs with evidence of both alleles is still the consernative thing to do, it may come at the cost of throwing out too much data.

Usually the input is expected to be the result of running a function to count reads in 10X data on a set of SNPS (\code{\link{getAllelicExpression}} or \code{\link{vartrixCnts}}).  However, this function will also accept the output of \code{\link{hetSNPsFromRNA}}.  This function has to calculate counts at all potential SNP sites as part of the estimation procedure, which it stores in the \code{@metadata} slot for later use.  This prevents having to calculate a pileup from the same files twice when estimating SNPs from the same RNA BAM files that counts are to be estimated from.
}
