% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSolutions.R
\name{plotSolutions}
\alias{plotSolutions}
\title{Plot summary of solutions}
\usage{
plotSolutions(
  fit,
  pointColour = "darkgrey",
  pointSize = 1,
  pointAlpha = 1,
  ...
)
}
\arguments{
\item{fit}{The output of \code{\link{inferInactiveX}}.}

\item{pointColour}{Colour of points.}

\item{pointSize}{Size of points.}

\item{pointAlpha}{Transparency of points.}

\item{...}{Passed to \code{\link{geom_hex}}.}
}
\value{
A ggplot object containing the plot.
}
\description{
The EM algorithm used by \code{\link{inferInactiveX}} is rerun with many random starting points (1,000 by default).  This function plots a summary of these solutions, showing the log likelihood of each solution (y-axis) against the fraction of cells allocated to maternal (x-axis).
}
\details{
When the genotyping is not known beforehand the designation of maternal or paternal is arbitrary and you should expect the solutions to be evenly distributed either side of 0.5 on the x-axis.
}
